Hotfix/execution bugs (#46)
* fix bug on ExectutionThread and BlockProgramLogic

* add a notice to show the error of connection

* fix a bug on ExecutionActivityBase

* remove is_connected data. it was uglily scattered in many places to notify an Android device whether it is already connected to a robot. make it worse, it caused some bugs

* add null checks for sensors

* remove useless Log.d

* add BluetoothCommunicatorTest

* add BlockProgramLogicTest

* also add mokito for unit testing

* add NxtControllerTest and fix bugs

* remove useless conversion

* update tests based on review comments

* change interfaces of BluetoothCommunicator / NxtController

* fix a little based on an interface change

* add the latest dexmaker by using JitPack.io

* fix a bug on a test

* avoid instantiate BluetoothDevice by specifying @NonNull

* comment out a test

* fix BluetoothCommunicator to throw NPE for null device

* comment out all tests related to PowerMock

* follow review comments

* revert build.gradle changes related to dexmaker

* add  modifier

+++ app/src/nxt/java/com/pileproject/drive/programming/visual/block/sequence/SetRightMotorSpeedBlock.java
@@ -57,7 +57,7 @@ public class SetRightMotorSpeedBlock extends SequenceBlockHasNumberText {
 
     @Override
     public int action(MachineController controller, ExecutionCondition condition) {
-        ((NxtController) controller).setMotorPower(NxtController.MotorKind.RightMotor, getValue().doubleValue());
+        ((NxtController) controller).setMotorPower(NxtController.MotorKind.RightMotor, getValue().intValue());
         return 0;
     }
